requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-Session==0.5.0
SQLAlchemy==2.0.36
PyMySQL==1.1.0
pytz==2023.3
python-dotenv==1.0.0